Refactor health components
